# quodom-web/api/openapi.yaml
openapi: 3.0.3
info:
  title: Quodom API
  version: 1.0.0
  description: |
    Contrato entre la app y el API. **Este archivo manda**: todo cambio de API
    empieza acá y después va el código. El servidor valida los pedidos contra
    este archivo, los tests validan las respuestas, y la app genera sus tipos
    con `npm run api-types`.
servers:
  - url: /
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: quodom_session
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    BadRequest:
      description: Pedido inválido o regla de negocio rota
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: Sin sesión o sesión inválida
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: No existe
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    TooMany:
      description: Demasiados intentos
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  parameters:
...
    QuodomId:
      name: id
      in: path
      required: true
      schema: { type: string }
  schemas:
    Error:
      type: object
      required: [res]
      properties:
        res: { type: boolean, enum: [false] }
        message: { type: string }
        error: { type: string }
        exp: { type: boolean }
      additionalProperties: true
...
    OkMessage:
      type: object
      required: [res, message]
      properties:
        res: { type: boolean }
        message: { type: string }
...
    BusquedaResult:
      type: object
      required: [id, nombre]
      properties:
        id: { type: integer }
        nombre: { type: string, nullable: true }
        descripcion: { type: string, nullable: true }
        imagen: { type: string, nullable: true }
        refreshImagen: { type: string, nullable: true }
        categoriaPadre: { type: integer, nullable: true }
...
    Direccion:
      type: object
      required: [id, userid]
      properties:
        id: { type: integer }
        userid: { type: string }
        provincia: { type: string, nullable: true }
        partido: { type: string, nullable: true }
        localidad: { type: string, nullable: true }
        direccion: { type: string, nullable: true }
        calle: { type: string, nullable: true }
        numero: { type: string, nullable: true }
        piso: { type: string, nullable: true }
        cp: { type: string, nullable: true }
        alias: { type: string, nullable: true }
        default: { anyOf: [{ type: boolean, nullable: true }, { type: integer, nullable: true }] }
        idprovincia: { type: string, nullable: true }
        idpartido: { type: string, nullable: true }
        observaciones: { type: string, nullable: true }
...
    Notificacion:
      type: object
      required: [id, titulo, texto]
      properties:
        id: { type: integer }
        userId: { type: string }
        tiponotificacion: { type: string }
        idquodom: { type: string }
        titulo: { type: string }
        texto: { type: string }
        enviada: { type: boolean, nullable: true }
        leida: { anyOf: [{ type: boolean, nullable: true }, { type: integer, nullable: true }] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CurrentUser:
      type: object
      required: [id, username, email]
      properties:
        id: { type: string }
        username: { type: string }
        email: { type: string }
        nombre: { type: string, nullable: true }
        apellido: { type: string, nullable: true }
        dni: { type: string, nullable: true }
        refreshFoto: { type: string, nullable: true }
        telefono: { type: string, nullable: true }
        codArea: { type: string, nullable: true }
    SigninResponse:
      type: object
      required: [res, id, username, email]
      description: El token NO viaja acá; queda en la cookie httpOnly quodom_session.
      properties:
        res: { type: boolean }
        id: { type: string }
        username: { type: string }
        email: { type: string }
        nombre: { type: string, nullable: true }
        apellido: { type: string, nullable: true }
        role: { type: string, nullable: true }
        refreshFoto: { type: string, nullable: true }
    InfoComprador:
      type: object
      properties:
        idquodom: { type: string }
        NombreComprador: { type: string, nullable: true }
        telefono: { type: string, nullable: true }
        email: { type: string, nullable: true }
        Direccion: { type: string, nullable: true }
        Provincia: { type: string, nullable: true }
        Localidad: { type: string, nullable: true }
        cp: { type: string, nullable: true }
        observaciones: { type: string, nullable: true }
paths:
...
  /busqueda:
    get:
      tags: [catalogo]
      parameters:
        - { name: b, in: query, required: false, schema: { type: string } }
      responses:
        '200':
          description: OK (lista vacía si no hay b)
          content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/BusquedaResult' } } } }
...
  /users:
    get:
      tags: [users]
      summary: Todos los usuarios (sólo admin)
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { type: array, items: { type: object, additionalProperties: true } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    put:
      tags: [users]
      summary: Editar el perfil propio
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                email: { type: string, format: email }
                nombre: { type: string }
                apellido: { type: string }
                dni: { type: string }
                codArea: { type: string }
                telefono: { type: string }
                password: { type: string, minLength: 6 }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OkMessage' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /users/current:
    get:
      tags: [users]
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/CurrentUser' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /users/currentFoto:
    get:
      tags: [users]
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  foto: { type: string, nullable: true }
                  refreshFoto: { type: string, nullable: true }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /users/dire:
    get:
      tags: [users]
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Direccion' } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /users/direcciondefault:
    get:
      tags: [users]
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200':
          description: La dirección por defecto, o la primera, o null
          content: { application/json: { schema: { oneOf: [{ $ref: '#/components/schemas/Direccion' }, { type: string, nullable: true, enum: [null] }] } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /users/infoComprador/{idquodom}:
    get:
      tags: [users]
      summary: Datos de contacto del dueño del Quodom (sólo el propio)
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: idquodom, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/InfoComprador' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /users/validateEmail/{token}:
    get:
      tags: [users]
      parameters:
        - { name: token, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: 'res false si el token no sirve', content: { application/json: { schema: { $ref: '#/components/schemas/OkMessage' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
  /users/validateReset/{token}:
    get:
      tags: [users]
      parameters:
        - { name: token, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: 'res false si el link venció o ya se usó'
          content:
            application/json:
              schema:
                type: object
                required: [res]
                properties:
                  res: { type: boolean }
                  message: { type: string }
        '429': { $ref: '#/components/responses/TooMany' }
  /users/{id}:
    get:
      tags: [users]
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/QuodomId' }]
      responses:
        '200':
          description: Sólo el propio usuario
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  username: { type: string }
                  email: { type: string }
                  nombre: { type: string, nullable: true }
                  apellido: { type: string, nullable: true }
        '401': { $ref: '#/components/responses/Unauthorized' }
    delete:
      tags: [users]
      summary: Borrar un usuario (sólo admin)
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/QuodomId' }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OkMessage' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /users/signin:
    post:
      tags: [users]
      summary: Inicia sesión y pone la cookie quodom_session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string, description: 'Usuario o e-mail' }
                password: { type: string }
      responses:
        '200':
          description: OK. Set-Cookie quodom_session (httpOnly, SameSite=Strict, 30 días)
          content: { application/json: { schema: { $ref: '#/components/schemas/SigninResponse' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/TooMany' }
...
  /users/signup:
    post:
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, nombre, password, codArea, telefono]
              properties:
                username: { type: string }
                email: { type: string, format: email }
                nombre: { type: string }
                apellido: { type: string }
                password: { type: string, minLength: 6 }
                dni: { type: string }
                codArea: { type: string }
                telefono: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [res, message]
                properties:
                  res: { type: boolean }
                  id: { type: string }
                  message: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/TooMany' }
  /users/reset:
    post:
      tags: [users]
      summary: Pide el link de blanqueo. Contesta igual exista o no la cuenta.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, description: 'E-mail o usuario' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OkMessage' } } } }
        '429': { $ref: '#/components/responses/TooMany' }
  /users/reenviar:
    post:
      tags: [users]
      summary: Reenvía el link de validación. Contesta igual exista o no la cuenta.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OkMessage' } } } }
        '429': { $ref: '#/components/responses/TooMany' }
  /users/changePass:
    post:
      tags: [users]
      summary: Cambia la clave con el token del link (sirve una sola vez)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password, token]
              properties:
                password: { type: string, minLength: 6 }
                token: { type: string }
      responses:
        '200': { description: 'res false si el link venció o ya se usó', content: { application/json: { schema: { $ref: '#/components/schemas/OkMessage' } } } }
        '429': { $ref: '#/components/responses/TooMany' }
  /users/cambiarFoto/{id}:
    put:
      tags: [users]
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/QuodomId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                foto: { type: string }
                refreshFoto: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  res: { type: boolean }
                  refresh: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /oper_notificaciones:
    get:
      tags: [oper_notificaciones]
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Notificacion' } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /hist_busquedas:
    get:
      tags: [hist_busquedas]
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200':
          description: Últimas búsquedas del usuario
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: integer }
                    valor: { type: string, nullable: true }
                    createdAt: { type: string, format: date-time }
        '401': { $ref: '#/components/responses/Unauthorized' }
...